revert - client throws IgniteClientAuthenticationException instead of AuthenticationException
